Move address resolution out of core and into app
