+++ tests/client/lwm2mclient.c
@@ -161,7 +161,7 @@ static void prv_output_servers(char * buffer,
         char s[INET6_ADDRSTRLEN];
 
         fprintf(stdout, "Server ID %d:\r\n", targetP->shortID);
-        fprintf(stdout, "\thost: \"%s\" port: %hu\r\n", targetP->host, targetP->port);
+        //fprintf(stdout, "\thost: \"%s\" port: %hu\r\n", targetP->host, targetP->port);
         fprintf(stdout, "\tstatus: ");
         switch(targetP->status)
         {
@@ -197,6 +197,53 @@ static void prv_change(char * buffer,
 
 syntax_error:
     fprintf(stdout, "Syntax error !");
+}
+
+static int prv_add_server(lwm2m_context_t * contextP,
+                          uint16_t shortID,
+                          char * host,
+                          uint16_t port,
+                          lwm2m_security_t * securityP)
+{
+    char portStr[6];
+    struct addrinfo hints;
+    struct addrinfo *servinfo = NULL;
+    struct addrinfo *p;
+    int sock;
+    int status = INTERNAL_SERVER_ERROR_5_00;
+    struct sockaddr *sa;
+    socklen_t sl;
+
+    memset(&hints, 0, sizeof(hints));
+    hints.ai_family = AF_UNSPEC;
+    hints.ai_socktype = SOCK_DGRAM;
+
+    if (0 >= sprintf(portStr, "%hu", port)) return INTERNAL_SERVER_ERROR_5_00;
+    if (0 != getaddrinfo(host, portStr, &hints, &servinfo) || servinfo == NULL) return NOT_FOUND_4_04;
+
+    // we test the various addresses
+    sock = -1;
+    for(p = servinfo ; p != NULL && sock == -1 ; p = p->ai_next)
+    {
+        sock = socket(p->ai_family, p->ai_socktype, p->ai_protocol);
+        if (sock >= 0)
+        {
+            sa = p->ai_addr;
+            sl = p->ai_addrlen;
+            if (-1 == connect(sock, p->ai_addr, p->ai_addrlen))
+            {
+                close(sock);
+                sock = -1;
+            }
+        }
+    }
+    if (sock >= 0)
+    {
+        status = lwm2m_add_server(contextP, shortID, sa, sl, securityP);
+        if (status != 0)
+            close(sock);
+    }
+    return status;
 }
 
 int main(int argc, char *argv[])
@@ -250,7 +297,7 @@ int main(int argc, char *argv[])
     signal(SIGINT, handle_sigint);
 
     memset(&security, 0, sizeof(lwm2m_security_t));
-    result = lwm2m_add_server(lwm2mH, 123, "::1", 5684, &security);
+    result = prv_add_server(lwm2mH, 123, "::1", 5684, &security);
     if (result != 0)
     {
         fprintf(stderr, "lwm2m_add_server() failed: 0x%X\r\n", result);
